[MODEL] change some property of table: City, Location, Feedback

+++ travel_api/Models/EF/Location.cs
@@ -19,7 +19,7 @@ namespace travel_api.Models.EF
         
         public decimal LocationLatitude { get; set; }
 
-        public float LocationRateAverage { get; set; }
+        public decimal LocationRateAverage { get; set; }
 
         public virtual ICollection<Post> Posts { get; set; }
